requirements.txt
streamlit
geopandas
streamlit-folium
matplotlib
sentinelhub
rioxarray
shapely
azure-keyvault-secrets
azure-identity
confuse
plotly
folium
mapclassify